remove template URL parameter from info tab
When navigating to the experiment info tab the template URL parameter persisted. This might be undesirable when the user navigates back and forth between the info and the workspace tab with a default template set. Without persisting template URL parameter the user is always navigated to the default template when switching from the info to the workspace tab.